convert to coin
