+++ app/build.gradle.kts
@@ -61,5 +61,5 @@ dependencies {
     implementation ("com.squareup.retrofit2:retrofit:2.9.0")
     implementation ("com.squareup.retrofit2:converter-gson:2.6.0")
     implementation ("com.squareup.okhttp3:logging-interceptor:3.12.6")
-
+    implementation("io.insert-koin:koin-android:4.0.4" )
 }
