+++ app/src/main/java/com/example/filmsearch/domain/Interactor.kt
@@ -13,7 +13,7 @@ import retrofit2.Response
 import com.example.filmsearch.utils.Conventer
 
 
-class Interactor(private val repo: MainRepository, private val retrofitService: TmdbApi) {
+class Interactor (private val repo: MainRepository, private val retrofitService: TmdbApi) {
     //В конструктор мы будм передавать коллбэк из вьюмоделе, чтобы реагировать на то, когда фильмы будут получены
     //и страницу, котороую нужно загрузить (это для пагинации)
     fun getFilmsFromApi(page: Int, callback: HomeFragmentViewModel.ApiCallback) {
